Revert "Move shared.types to type - conflict with stdlib types under docker/py3.12"
This reverts commit 72c48698a0a06f53c97fb4b8c36e31e004910f12.
